[JHBuild] Replace download.gnome.org for GitHub mirror or gitlab.gnome.org
Many times downloads of 'libsoup', 'glib' or 'glib-networking' fail unexpectedly.
This might be due to a problem on server side. Apparently, GitHub mirrors are
more reliably. For those GNOME repositories which do not have a GitHub mirror,
like 'atk' or 'at-spi-atk', I changed the type of repository from tarball to git.