[PATCH] WeakSet should count its heap allocations https://bugs.webkit.org/show_bug.cgi?id=295072
Reviewed by NOBODY (OOPS!).

Long running pages that do not allocate, but have a RAF callback
can see their memory usage slowly increase forever. This is caused
by WeakBlock allocations. Without other kinds of allocations,
the collector thinks that the footprint isn't growing.

We notify the collector about these allocations now. Notably, it is
very important that we do not GC while allocating a Weak<T>, because
we construct them all over the place in the middle of inserting
into WeakMap. `didAllocate` does not trigger GC on its own, so this
patch does not change this behaviour.