REGRESSION(299386@main): [GLib] Do not cancel the connection when stopping the write socket monitor https://bugs.webkit.org/show_bug.cgi?id=298192
Reviewed by Adrian Perez de Castro.

GSocketMonitor internally uses a GCancellable to cancel the source when
the monitor is stopped. ConnectionGLib uses a global GCancellable to
cancel the async operations, so in 299386@main we added the possibility
to pass an external GCancellable to GSocketMonitor. The problem is that
ConnectionGLib starts and stop the write socket monitor on demand, so
after the first time it's used the connection cancellable is cancelled
when stopped causing the connection to be closed unexpectedly. We should
only cancel the GCanbablle when stopping the socket monitor if it's not
a external one.

Originally-landed-as: 299418@main (66771d91fe99). https://bugs.webkit.org/show_bug.cgi?id=298192